hackathon-18march-boilerplate/personal_assistant/requirements.txt
orca-platform-sdk-ui
uvicorn
fastapi
anthropic
python-dotenv
